Review AzureDevOpsRestClient
